Move `FastRandom` to `LegacyRandom` in `osu.Game` project
